signTxCVoteRegistration: update for Stax
